\name{read_dashboard}
\alias{read_dashboard}
\title{Read a BlinkDash project}
\usage{
read_dashboard(project)
}
\arguments{
\item{project}{Project directory or \code{dashboard.blink.yml}/\code{dashboard.blink.json} file.}
}
\value{
A dashboard manifest list.
}
\description{
Reads a BlinkDash project from YAML or JSON.
}
